# openapi.yaml
openapi: 3.0.3
info:
  title: CIST Portal Scraping API
  description: CIST ポータルサイトからスクレイピングしたデータを取得する API の仕様書です。
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: ローカル開発サーバー
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      security:
        - {}
      summary: 招待コード付きユーザー登録
      description: ユーザー情報とCIST-Accountを保存し、メール認証なしでJWTとRefreshトークンを発行します。
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /api/auth/login:
    post:
      security:
        - {}
      summary: ユーザー名またはメールアドレスでログイン
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /api/auth/refresh:
    post:
      security:
        - {}
      summary: Refreshトークンをローテーションして再発行
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 再発行成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
...
  /api/auth/logout:
    post:
      summary: Refreshトークンを失効してログアウト
      operationId: logoutUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: ログアウト成功
...
  /api/auth/passkeys/register/verify:
    post:
      summary: Passkey登録検証
      operationId: verifyPasskeyRegistration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Passkey登録成功
  /api/auth/passkeys/login/options:
    post:
      security:
        - {}
      summary: Passkeyログインオプション生成
      operationId: createPasskeyLoginOptions
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                  description: 任意のユーザー名またはメールアドレス
      responses:
        '200':
          description: WebAuthn authentication options
  /api/auth/passkeys/login/verify:
    post:
      security:
        - {}
      summary: Passkeyログイン検証
      operationId: verifyPasskeyLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /api/auth/passkeys/{credentialId}:
    delete:
      summary: 登録済みPasskey削除
      operationId: deletePasskey
      parameters:
        - name: credentialId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 削除成功
  /api/admin/invites:
    post:
      summary: 招待コード作成
      operationId: createInvite
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInDays:
                  type: integer
                  minimum: 1
      responses:
        '201':
          description: 作成成功。招待コードはこのレスポンスでのみ返ります。
    get:
      summary: 招待一覧
      operationId: listInvites
      responses:
        '200':
          description: 取得成功
  /api/admin/invites/{inviteId}:
    delete:
      summary: 招待コード失効
      operationId: revokeInvite
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 失効成功
  /api/mypage:
    get:
      summary: マイページデータ一括取得
      description: 新着情報、レポートサマリー、未提出レポート、振り返りへの返信、時間割などのマイページデータを一括で取得します。
      operationId: getMyPage
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyPageData'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mypage-summaries:
    get:
      summary: マイページサマリー取得
      description: マイページの新着情報サマリーと未提出レポートのサマリーを取得します。
      operationId: getMyPageSummaries
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - new_information
                  - unsubmitted_report_summary
                properties:
                  new_information:
                    $ref: '#/components/schemas/MyPageNewInformationSummary'
                  unsubmitted_report_summary:
                    $ref: '#/components/schemas/UnsubmittedReportSummary'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mypage-new-information:
    get:
      summary: マイページ新着情報サマリー取得
      description: 連絡通知、スケジュール、アンケートなどの新着未処理件数のサマリーを取得します。
      operationId: getMyPageNewInformation
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyPageNewInformationSummary'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/unsubmitted-report-summary:
    get:
      summary: 未提出レポートサマリー取得
      description: 期限が1週間以内の未提出レポート数と、全体の未提出レポート数のサマリーを取得します。
      operationId: getUnsubmittedReportSummary
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsubmittedReportSummary'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/unsubmitted-reports:
    get:
      summary: 未提出レポート一覧取得
      description: 未提出レポートの詳細一覧を取得します。
      operationId: getUnsubmittedReports
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnsubmittedReport'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reflection-replies:
    get:
      summary: 振り返り（授業コメントなど）の返信一覧取得
      description: 授業等の振り返りに対して教員などから送信された返信の一覧を取得します。
      operationId: getReflectionReplies
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReflectionReply'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/timetable:
    get:
      summary: 時間割一覧取得
      description: 今日または指定期間の時間割情報を取得します。
      operationId: getTimetable
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimetableEntry'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/received-office-memos:
    get:
      summary: 受信連絡通知一覧取得
      description: フィルタ、検索キーワード、カテゴリ、ページを指定して受信した連絡通知の一覧をページネーション付きで取得します。
      operationId: getReceivedOfficeMemos
      parameters:
        - name: filter
          in: query
          description: '抽出フィルタ ("all", "unread", "read", "star")'
          required: true
          schema:
            $ref: '#/components/schemas/OfficeMemoFilter'
        - name: searchKeyword
          in: query
          description: 検索用キーワード
          required: false
          schema:
            type: string
            default: ""
        - name: c_filter
          in: query
          description: カテゴリによる抽出フィルタ
          required: false
          schema:
            type: string
            default: "c_all"
        - name: page
          in: query
          description: 取得ページ番号（1から開始）
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReceivedOfficeMemos'
        '400':
          description: リクエスト不正（パラメータエラーなど）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pending-appointments-summary:
    get:
      summary: 保留中のアポイントメントサマリー取得
      description: 保留中の面談や予定予約の件数および保留中アイテムの有無を取得します。
      operationId: getPendingAppointmentsSummary
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingAppointmentSummary'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/monthly-schedule:
    get:
      summary: 月間スケジュール取得
      description: スケジュールに登録されているイベント月間の一覧を取得します。
      operationId: getMonthlySchedule
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlySchedule'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/undone-questionnaires:
    get:
      summary: 未回答アンケート一覧取得
      description: まだ回答していないアクティブなアンケートの一覧を取得します。
      operationId: getUndoneQuestionnaires
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UndoneQuestionnaire'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/received-office-memo-details:
    get:
      summary: 受信連絡詳細取得
      description: 指定された `officeMemoId` を持つ連絡通知の本文や添付ファイル等を含む詳細情報を取得します。
      operationId: getReceivedOfficeMemoDetails
      parameters:
        - name: officeMemoId
          in: query
          description: 連絡通知の一意なID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceivedOfficeMemoDetail'
        '400':
          description: リクエスト不正（ID未指定など）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/courses-for-user:
    get:
      summary: ユーザ履修講義一覧取得
      description: ユーザが履修中、または関連する講義（コース）の一覧を取得します。
      operationId: getCoursesForUser
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseSummary'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/course-lecture-details:
    get:
      summary: 講義詳細一覧取得
      description: 指定された講義 (`courseId`) の全セッション、または特定セッション of 講義詳細情報を取得します。
      operationId: getCourseLectureDetails
      parameters:
        - name: courseId
          in: query
          description: 講義（コース）の一意なID
          required: false
          schema:
            type: string
        - name: session
          in: query
          description: セッション回数（正の整数）
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseLectureDetail'
        '400':
          description: リクエスト不正（セッションパラメータ不正など）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/course-lecture-attendance:
    post:
      summary: 講義出席登録
      description: 指定した講義回IDと出席パスワードで出席登録を行います。
      operationId: registerCourseLectureAttendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseLectureAttendanceRequest'
      responses:
        '200':
          description: 出席登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseLectureAttendanceResult'
        '400':
          description: リクエスト不正、またはポータル側の出席登録エラー
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CourseLectureAttendanceResult'
                  - $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/distribution-pdf-urls:
    get:
      summary: 配布PDF of URLリスト取得
      description: 学生向け等にポータルで配布されているPDFファイルのジャンル別URL一覧を取得します。
      operationId: getDistributionPdfUrls
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistributionPdfGroup'
        '400':
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - inviteCode
        - username
        - email
        - password
        - cistAccount
      properties:
        inviteCode:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        cistAccount:
          type: object
          required:
            - userId
            - password
          properties:
            userId:
              type: string
              description: CISTポータルのログインUserID
            password:
              type: string
              format: password
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: ユーザー名またはメールアドレス
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthTokenResponse:
      type: object
      required:
        - user
        - accessToken
        - refreshToken
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
        accessToken:
          type: string
        refreshToken:
          type: string
...
    UnsubmittedReportSummary:
      type: object
      required:
        - due_within_week_count
        - total_count
      properties:
        due_within_week_count:
          type: integer
          description: 提出期限が1週間以内の未提出レポート数
        total_count:
          type: integer
          description: 総未提出レポート数
...
    OfficeMemoFilter:
      type: string
      enum:
        - all
        - unread
        - read
        - star
      description: 連絡通知のフィルタタイプ
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ